scala3 test: fixed the conversion in the AnyMatchersSpec

+++ tests/src/test/scala/org/specs2/matcher/AnyMatchersSpec.scala
@@ -132,4 +132,4 @@ case class Hello() { override def toString = "hello" }
 
 trait Conversions:
   given Conversion[Int, Long] with
-    def apply(n: Int): Long = n
+    def apply(n: Int): Long = n.toLong
